refactor: adapt github.com/bassosimone/clip v0.8.0
This diff adapts existing github.com/bassosimone/clip v0.8.0 code
and specifically the nparser package to create a new package.

The main change since the original is that we don't want anymore
the full command line, rather just the arguments.

Apart from this, we have also ensured that the early options stop
when they see a separator, which is good because otherwise one
cannot combine parsers using different conventions.

Also, we have simplified and rationalized the existing tests.