fix: handle early prefixes uniformly
When we adapted github.com/bassosimone/clip code, we also modified it
to scan before processing early options. We implemented this change
such that the following case is handled correctly:

    tool -abc --verbose -- subtool --help

Specifically, if `--` is a separator, scanning ensures that we stop
looking for options and thus we do not classify the `--help` token as
an early option, rather, we classify it as a positional.

However, this change surfaced an additional issue, which is the original
issue that prompted authoring this commit in the first place. When
short options are only ever used for early options, its prefixes are
not added to the map of prefixes, therefore:

    weekly ls -h

produces the following error:

    weekly init: too many positional arguments: expected at most 0, got 1

To fix this issue, we created a regression test and then modified the
`config.go` code to register prefixes used for early options.

Then, we created a regression test for the opposite case (i.e., only
short options registered and `--help` being passed on the command line)
and this test was succeeding. Turns out that there was a bug in the
`early.go` code, which was using string matching to find early options
rather than matching the prefix and name independently. We also
fixed this bug.

Then, reviewing the code, we noticed that `doparse.go` has a place
where it panices when a prefix is neither groupable or standalone, but,
due to the recent changes, this means that:

    weekly ls -x

would now panic if `-` is only used for the early `-h` option.

So, we created a test case for that and modified the code so that
when we find a prefix associated with early options only it's still
an option prefix that exists, so we turn that into an unknown
option error, and `weekly ls -x` now produces:

    unknown option: -x

as an error.